+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * You can now add the enum argument `cant_skip` to the end of a pause statement
   to ignore advances until the pause ends
+* You can now add the enum argument `no_wait` to the end of an add or say
+  statement to auto advance the dialog
 
 ### Changed
 * The compiler no longer uses a tokenizer except when parsing expressions